+++ resources/js/bootstrap/stores/plugin.js
@@ -18,6 +18,10 @@ import {
 } from '@/helpers/plugins.js';
 
 import {
+    availableDynalots,
+} from '@/helpers/plugin/plugin-dynalot.js';
+
+import {
     only,
     slugify,
 } from '@/helpers/helpers.js';
@@ -45,7 +49,18 @@ export const usePluginStore = defineStore('plugin', {
             tab: [],
             tools: [],
             settings: [],
+            dataModelOptions: [],
         },
+        /**
+         * Dynalots are similar to slots, only that they don't have a single rigid position in the
+         * UI but provide components to be rendered under specific conditions, 
+         * e.g. on the entity detail page, a various amount of tabs can be added dependent on the
+         * plugin data provided for the subscription.
+         */
+        registeredDynalots: availableDynalots().reduce((acc, topic) => {
+            acc[topic] = [];
+            return acc;
+        }, {}),
     }),
     actions: {
         add(plugin) {
@@ -144,6 +159,31 @@ export const usePluginStore = defineStore('plugin', {
             }
 
             return this.registeredSlots[slotName] ?? [];
+        },
+        getDynalotItems(slot) {
+            console.log(this.registeredDynalots);
+
+            if(!this.registeredDynalots[slot]) {
+                console.error('Dynalot does not exist', slot);
+                return [];
+            }
+
+            return this.registeredDynalots[slot] ?? [];
+        },
+        updateDynalots(slot, context) {
+            const dynalots = this.registeredDynalots[slot];
+            if(!dynalots) {
+                console.error('Dynalot does not exist', slot);
+                return;
+            }
+
+            dynalots.forEach(dynalot => {
+                try {
+                    dynalot.update(context)
+                } catch(e) {
+                    console.error('Error updating dynalot', dynalot, e);
+                }
+            });
         },
         async publishScript(plugin) {
             console.log('Publishing script for plugin', plugin);
@@ -244,6 +284,18 @@ export const usePluginStore = defineStore('plugin', {
             }
 
             this.registeredSlots[slot].push(data);
+        },
+        registerDynalot(data) {
+            if(!data.slot || !data.of || !data.slot) {
+                console.error('Plugin dynalot is missing slot information', data);
+            } else {
+                if(!this.registeredDynalots[data.slot]) {
+                    console.error('Plugin dynalot slot is not supported', data.slot);
+                    return;
+                }
+
+                this.registeredDynalots[data.slot].push(data);
+            }
         },
         reset() {
             this.plugins = [];